# openapi.yaml
openapi: 3.1.0
info:
  title: "Who Am I API"
  version: "v1"
  description: "Tiny API that prints OS and HTTP request info as output"
...
paths:
...
  /apispec:
    get:
      summary: Returns the whoami API specification
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum:
              - json
              - yaml
      response:
        '200':
          description: Whoami API specification
          content:
            text/plain:
            scheme:
              type: string
              format: binary
...
  /data:
    get:
      summary: Returns a data response with the specified size parameter
      parameters:
        - in: query
          name: size
          schema:
            type: integer
        - in: query
          name: unit
          schema:
            type: string
            enum:
              - kb
              - mb
              - gb
              - tb
        - in: query
          name: attachment
          schema:
            type: boolean
            allowEmptyValue: true
      response:
        '200':
          description: OK
        '500':
          description: Error while copying content
...